+++ src/stronghold/mod.rs
@@ -48,7 +48,6 @@
 
 mod common;
 mod db;
-mod encryption;
 mod secret;
 
 use std::{